Rename this package to elasticsearch-query (resolves #7)
