+++ Makefile
@@ -1,25 +1,26 @@
-coverage_options = --include='wikia_common_kibana.py' --omit='test/*'
+coverage_options = --include='elasticsearch_query.py' --omit='test/*'
 
 install:
 	pip install -e .[dev]
 
 test:
-	pytest -v
+	pytest -vv
 
 coverage:
 	rm -f .coverage*
 	rm -rf htmlcov/*
-	coverage run -p -m pytest -v
+	coverage run -p -m pytest -vv
 	coverage combine
 	coverage html -d htmlcov $(coverage_options)
 	coverage xml -i
 	coverage report $(coverage_options)
 
 lint:
-	pylint wikia_common_kibana.py
+	pylint elasticsearch_query.py
 
 publish:
 	# run git tag -a v0.0.0 before running make publish
-	python setup.py sdist upload -r pypi
+	python setup.py sdist
+	twine upload dist/*
 
 .PHONY: test
